+++ packages/pangea-multipass/pangea_multipass/core.py
@@ -53,6 +53,7 @@ class PangeaMetadataKeys(str, enum.Enum):
     GITLAB_REPOSITORY_NAME = f"{_PANGEA_METADATA_KEY_PREFIX}gitlab_repository_name"
     GITLAB_REPOSITORY_NAMESPACE_WITH_PATH = f"{_PANGEA_METADATA_KEY_PREFIX}gitlab_repository_namespace_with_path"
     DROPBOX_ID = f"{_PANGEA_METADATA_KEY_PREFIX}dropbox_id"
+    DROPBOX_PATH = f"{_PANGEA_METADATA_KEY_PREFIX}path"
     DROPBOX_FILE_PATH = f"{_PANGEA_METADATA_KEY_PREFIX}file_path"