+++ packages/pangea-multipass/pangea_multipass/github_reader.py
@@ -1,6 +1,4 @@
-from typing import Any, List
-
-import requests
+from typing import Any, List, Optional
 
 from .core import MultipassDocument, PangeaMetadataKeys, PangeaMetadataValues, generate_id
 from .sources.github import GitHubAPI
@@ -10,33 +8,41 @@ class GitHubReader:
     _token: str
     """GitHub personal access token"""
 
+    _current_file: int = 0
+    _repo_files: Optional[List[dict]] = None
+    _current_repository: dict = {}
+
     def __init__(self, token: str):
         self._token = token
+        self._restart()
 
     def load_data(
         self,
     ) -> List[MultipassDocument]:
-        # Authenticate
-        headers = GitHubAPI.get_auth_headers(self._token)
+        """
+        Load all the data from the repositories
+        This will read all the files from all the repositories the token has access to.
+        This process is blocking and can take a long time. If working with a large number of repositories,
+        consider using the read_repo_files method.
+        """
+        documents: List[MultipassDocument] = []
 
         # Get repositories
         repos = GitHubAPI.get_user_repos(self._token)
-
-        documents: List[MultipassDocument] = []
 
         for repo in repos:
             owner = repo["owner"]["login"]
             repo_name = repo["name"]
 
             # Get all files recursively
-            files = self._get_repo_files(headers, owner, repo_name)
+            files = GitHubAPI.get_repo_files(self._token, owner, repo_name)
 
             for file in files:
                 file_path = file["path"]
-                download_url = file["download_url"]
+                download_url = file["url"]
 
                 # Fetch the file content
-                content = self._download_file_content(headers, download_url)
+                content = GitHubAPI.download_file_content(self._token, download_url)
 
                 # Create metadata
                 metadata: dict[str, Any] = {
@@ -53,30 +59,72 @@ class GitHubReader:
 
         return documents
 
-    def _get_repo_files(self, headers: dict[str, Any], owner: str, repo: str, path: str = "") -> List[dict[str, Any]]:
-        """Recursively fetch all files and directories in a repository."""
+    def read_repo_files(self, repository: dict, page_size: int = 100) -> List[MultipassDocument]:
+        """
+        Read files from a given repository
+        If the repository is different from the last one, it will restart the reader.
+        If the repository is the same, it will continue reading from the last file.
+        """
+        documents: List[MultipassDocument] = []
 
-        url = f"https://api.github.com/repos/{owner}/{repo}/contents/{path}"
-        response = requests.get(url, headers=headers)
+        self._read_repo_files_checks(repository)
+        if self._repo_files is None:
+            return documents
 
-        if response.status_code == 200:
-            items = response.json()
-            files: List[dict[str, Any]] = []
-            for item in items:
-                if item["type"] == "file":
-                    files.append(item)
-                elif item["type"] == "dir":
-                    files.extend(self._get_repo_files(headers, owner, repo, item["path"]))
-            return files
-        elif response.status_code == 404:
-            return []
-        else:
-            raise Exception(f"Error fetching files for repository '{repo}': {response.json()}")
+        owner = self._current_repository["owner"]["login"]
+        repo_name = self._current_repository["name"]
 
-    def _download_file_content(self, headers: dict[str, Any], url: str) -> str:
-        """Download the content of a file from GitHub."""
-        response = requests.get(url, headers=headers)
-        if response.status_code == 200:
-            return str(response.content)
-        else:
-            raise Exception(f"Error downloading file: {response.json()}")
+        i = 0
+        while i < page_size and self._current_file < len(self._repo_files):
+            file = self._repo_files[self._current_file]
+            i += 1
+            self._current_file += 1
+
+            file_path = file["path"]
+            download_url = file["url"]
+
+            # Fetch the file content
+            content = GitHubAPI.download_file_content(self._token, download_url)
+
+            # Create metadata
+            metadata: dict[str, Any] = {
+                PangeaMetadataKeys.GITHUB_REPOSITORY_NAME: repo_name,
+                PangeaMetadataKeys.GITHUB_REPOSITORY_OWNER: owner,
+                PangeaMetadataKeys.FILE_PATH: file_path,
+                PangeaMetadataKeys.FILE_NAME: file_path,
+                PangeaMetadataKeys.DATA_SOURCE: PangeaMetadataValues.DATA_SOURCE_GITHUB,
+                PangeaMetadataKeys.GITHUB_REPOSITORY_OWNER_AND_NAME: (owner, repo_name),
+            }
+
+            doc = MultipassDocument(id=generate_id(), content=content, metadata=metadata)
+            documents.append(doc)
+
+        return documents
+
+    def get_repos(self) -> List[dict]:
+        """Get all the repositories the token has access to"""
+        return GitHubAPI.get_user_repos(self._token)
+
+    @property
+    def has_more_files(self):
+        """Check if there are more files to read"""
+        return self._repo_files is not None and self._current_file < len(self._repo_files)
+
+    def _restart(self) -> None:
+        self._current_file = 0
+        self._repo_files = None
+        self._current_repository = {}
+
+    def _read_repo_files_checks(self, repository: dict) -> None:
+        current_repo_id = self._current_repository.get("id", None)
+        new_repo_id = repository.get("id", None)
+
+        if current_repo_id is None or current_repo_id != new_repo_id:
+            self._restart()
+            self._current_repository = repository
+
+        owner = self._current_repository["owner"]["login"]
+        repo_name = self._current_repository["name"]
+
+        if self._repo_files is None:
+            self._repo_files = GitHubAPI.get_repo_files(self._token, owner, repo_name)