+++ packages/pangea-multipass/CHANGELOG.md
@@ -11,11 +11,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - GitLabReader and GitLabProcessor
 - Dropbox reader and processor
+- Pagination support on GitHubReader
+- Pagination support on SlackReader
 
 ### Fixed
 
 - Handle null fields on issues in JiraME
 - Handle trailing slash in Jira URL
+- GitLabProcessor `get_filter()`
 
 
 ## 0.2.0 - 2025-01-15